P0-20: fix all clippy warnings across workspace
Resolve all clippy lints so `cargo clippy --workspace -- -D warnings`
passes cleanly. Fixes include:

- fastclaw-mcp: type aliases for complex handler maps, Box large enum
  variant, while-let loop, strip_prefix, pass Arc by ref, generalize
  HashMap hasher, remove identity map
- fastclaw-agent: #[allow] for dead code and too_many_arguments,
  merge identical if-branches, iterator enumerate instead of index,
  strip_prefix, Default impls, redundant closures/derefs (auto-fixed)
- fastclaw-gateway: #[allow] for too_many_arguments, remove needless
  struct update, auto-fixed minor lints
- fastclaw-cli: matches! macro, auto-fixed format args and closures
- fastclaw-app: #[allow] for too_many_arguments and type_complexity,
  remove redundant binding
- fastclaw-feishu: derive Default, rename from_str to parse, fix
  match-as-equality, remove useless .into(), type_complexity allow

Verified: cargo build --workspace, cargo build -p fastclaw-gateway
--no-default-features, --all-features, and cargo clippy --workspace
-- -D warnings all pass.

Made-with: Cursor

+++ crates/fastclaw-gateway/src/consolidation.rs
@@ -149,6 +149,7 @@ fn build_compact_context(messages: &[ChatMessage]) -> String {
     buf
 }
 
+#[allow(clippy::too_many_arguments)]
 async fn run_consolidation(
     llm: &dyn LlmProvider,
     model: &str,
@@ -239,7 +240,7 @@ async fn run_consolidation(
                 predicate: pred.clone(),
                 object: obj.clone(),
                 category: FactCategory::UserPreference.as_str().to_string(),
-                confidence: importance.clamp(0.0, 1.0) as f32,
+                confidence: importance.clamp(0.0, 1.0),
                 source_session: Some(session_id.to_string()),
                 created_at: now.clone(),
                 updated_at: now,
